# Alerting: Improve performance of matching captures