--- a/pkg/services/ngalert/eval/eval.go
+++ b/pkg/services/ngalert/eval/eval.go
@@ -303,14 +303,19 @@ type NumberValueCapture struct {
 
 func queryDataResponseToExecutionResults(c models.Condition, execResp *backend.QueryDataResponse) ExecutionResults {
 	// captures contains the values of all instant queries and expressions for each dimension
-	captures := make(map[data.Fingerprint][]NumberValueCapture, len(execResp.Responses))
+	captures := make(map[string]map[data.Fingerprint]NumberValueCapture)
 	captureFn := func(refID string, labels data.Labels, value *float64) {
+		m := captures[refID]
+		if m == nil {
+			m = make(map[data.Fingerprint]NumberValueCapture)
+		}
 		fp := labels.Fingerprint()
-		captures[fp] = append(captures[fp], NumberValueCapture{
+		m[fp] = NumberValueCapture{
 			Var:    refID,
 			Value:  value,
 			Labels: labels.Copy(),
-		})
+		}
+		captures[refID] = m
 	}
 
 	// datasourceUIDsForRefIDs is a short-lived lookup table of RefID to DatasourceUID
@@ -382,24 +387,23 @@ func queryDataResponseToExecutionResults(c models.Condition, execResp *backend.Q
 			theseLabels := frame.Fields[0].Labels
 			fp := theseLabels.Fingerprint()
 
-			// First look for a capture whose labels are an exact match
-			if groupedCaps, ok := captures[fp]; ok {
-				if frame.Meta.Custom == nil {
-					frame.Meta.Custom = []NumberValueCapture{}
-				}
-				frame.Meta.Custom = append(frame.Meta.Custom.([]NumberValueCapture), groupedCaps...)
-			} else {
-				// If no exact match was found, look for captures whose labels are either subsets
-				// or supersets
-				for _, groupedCaps := range captures {
-					if len(groupedCaps) > 0 {
-						firstCap := groupedCaps[0]
+			for _, fps := range captures {
+				// First look for a capture whose labels are an exact match
+				if v, ok := fps[fp]; ok {
+					if frame.Meta.Custom == nil {
+						frame.Meta.Custom = []NumberValueCapture{}
+					}
+					frame.Meta.Custom = append(frame.Meta.Custom.([]NumberValueCapture), v)
+				} else {
+					// If no exact match was found, look for captures whose labels are either subsets
+					// or supersets
+					for _, v := range fps {
 						// matching labels are equal labels, or when one set of labels includes the labels of the other.
-						if theseLabels.Equals(firstCap.Labels) || theseLabels.Contains(firstCap.Labels) || firstCap.Labels.Contains(theseLabels) {
+						if theseLabels.Equals(v.Labels) || theseLabels.Contains(v.Labels) || v.Labels.Contains(theseLabels) {
 							if frame.Meta.Custom == nil {
 								frame.Meta.Custom = []NumberValueCapture{}
 							}
-							frame.Meta.Custom = append(frame.Meta.Custom.([]NumberValueCapture), groupedCaps...)
+							frame.Meta.Custom = append(frame.Meta.Custom.([]NumberValueCapture), v)
 						}
 					}
 				}